Add "to have" at end of first sentence.
So I would extend this section probably. At least by a few paragraphs. It seems too short/abrupt to have it's own section otherwise.

I'm not sure if you go into this later, but this may be a good place to insert the problem solving steps we went over early on in algorithms class.